requirements.txt
algoliasearch>=4,<5
beautifulsoup4
lxml
python-dotenv
requests